xilinx-sdfec: Interrupt example
The example processor application is used to demonstrate
how to configure the SD-FEC IP instance(s) for interrupt
usage with the Linux SDFEC driver.
This application is for ZCU111 board only.
To run the application do “sdfec-interrupts”.

The Interrupt example provides APP recipes for a petalinux
project and can be built inside petalinux project only.
The README explains the process how to apply recepies to
a petalinux project.

Signed-off-by: Dragan Cvetic <dragan.cvetic@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>