xilinx-sdfec: Multi LDPC code example
The example application is used to set up and control
the BER test HW and demonstrates how to configure and
use the SD-FEC IP instance(s) using the Linux SDFEC
driver for three different LDPC parameter types.
This application is for ZCU111 board only.
To run the application do "sdfec-multi-ldpc-codes”

Signed-off-by: Dragan Cvetic <dragan.cvetic@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>