xilinx-sdfec: Remove dependency on meta-petalinux
The library "sdfecusrintf" is meta-petalinux property.
The library is a small wrapper for SDFEC system calls.
The sdfec examples depend on "sdfecusrintf" library.
Migrating the library files into examples repository
removes need for the library in meta-petalinux.

This patch set correct file modes, also removes .bb
files as they are not needed here.

Signed-off-by: Dragan Cvetic <dragan.cvetic@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>