xilinx-sdfec: Remove libmetal dependency
Remove a dependency on libmetal from SDFEC example code.
The libmetal is not in sync with mainline kernel
which makes the SDFEC example APP fails on libmetali
operations if built as a part of mainline kernel.
That makes imposible to validate on HW the SDFEC code
changes suggested by review requests.